auth: consolidate auth code in auth.rs; isolate the native token mechanism
Auth logic was split awkwardly — the auth types and the Auth facade lived in
auth.rs, but ~250 lines of the actual machinery (AuthState/AuthHeader and the
impl Rest auth methods) sat in rest.rs. Move it so auth.rs owns the durable
auth core and rest.rs keeps only the client-owned state field and the generic
request pipeline. Behaviour-preserving verbatim move (no logic/signature
changes).

Moved into auth.rs (as impl Rest blocks): AuthState/AuthHeader, acquire_token,
get_auth_header, fetch_token_from_url (the authURL path — raw HttpClient, no
Ably pipeline), auth_config[_with], effective_token_params,
check_client_id_compat, token_auth_in_effect, adjusted_now_ms,
invalidate_cached_token. The Key TYPE, the Auth facade, and revoke_tokens
(same for all token types) stay in auth.rs.

New token_request.rs isolates the deprecatable Ably-native-token mechanism:
exchange_token_request + token_request_timestamp (impl Rest) and Key::sign/
sign_with_timestamp (impl Key). When native tokens give way to JWT, this file
and the two acquire_token branches that call it are a bounded excision; basic
auth, authURL, callbacks, revocation, and the state/header machinery are
untouched. rest.rs keeps Mutex<auth::AuthState> (client-owned state).

Only pub(crate) visibility widening (AuthHeader::value, do_request_internal);
nothing widened beyond crate. Suite unchanged: unit 1258/0/15, live
integration 91/0, clippy + fmt clean.
