+++ src/auth.rs
@@ -3,13 +3,11 @@ use std::pin::Pin;
 use std::sync::Arc;
 
 use chrono::{DateTime, Utc};
-use hmac::{Hmac, Mac};
 use serde::{Deserialize, Serialize};
-use sha2::Sha256;
 
 use crate::error::{ErrorCode, ErrorInfo, Result};
-
-type HmacSha256 = Hmac<Sha256>;
+use crate::http_client::HttpRequest;
+use crate::rest::Rest;
 
 #[derive(Clone, Serialize, Deserialize)]
 pub struct Key {
@@ -28,64 +26,6 @@ impl Key {
             name: parts[0].to_string(),
             value: parts[1].to_string(),
         })
-    }
-
-    /// Sign a TokenRequest (RSA9). Fields not specified in `params` are
-    /// omitted from the request (RSA5/RSA6) and signed as empty strings.
-    /// `timestamp_ms` must already be resolved (local or server time, RSA9d).
-    pub(crate) fn sign_with_timestamp(
-        &self,
-        params: &TokenParams,
-        timestamp_ms: i64,
-    ) -> Result<TokenRequest> {
-        // RSA5/RSA6: ttl and capability are signed as empty strings and
-        // omitted from the TokenRequest when unspecified, so Ably applies
-        // the key defaults server-side.
-        let ttl_text = params.ttl.map(|t| t.to_string()).unwrap_or_default();
-        let capability = params.capability.as_deref().unwrap_or("");
-        let client_id = params.client_id.as_deref().unwrap_or("");
-
-        let nonce = match &params.nonce {
-            Some(n) => n.clone(),
-            None => {
-                let mut nonce_bytes = [0u8; 16];
-                rand::RngCore::fill_bytes(&mut rand::thread_rng(), &mut nonce_bytes);
-                base64::encode(nonce_bytes)
-            }
-        };
-
-        let sign_text = format!(
-            "{}\n{}\n{}\n{}\n{}\n{}\n",
-            self.name, ttl_text, capability, client_id, timestamp_ms, nonce,
-        );
-
-        let mut mac = HmacSha256::new_from_slice(self.value.as_bytes())?;
-        mac.update(sign_text.as_bytes());
-        let mac_b64 = base64::encode(mac.finalize().into_bytes());
-
-        Ok(TokenRequest {
-            key_name: self.name.clone(),
-            ttl: params.ttl,
-            capability: params.capability.clone(),
-            client_id: if client_id.is_empty() {
-                None
-            } else {
-                Some(client_id.to_string())
-            },
-            timestamp: Some(timestamp_ms),
-            nonce,
-            mac: mac_b64,
-        })
-    }
-
-    /// Sign a TokenRequest using the local clock (or the explicit timestamp
-    /// in `params`).
-    pub fn sign(&self, params: &TokenParams) -> Result<TokenRequest> {
-        let timestamp = params
-            .timestamp
-            .map(|t| t.timestamp_millis())
-            .unwrap_or_else(|| Utc::now().timestamp_millis());
-        self.sign_with_timestamp(params, timestamp)
     }
 }
 
@@ -591,5 +531,383 @@ impl AuthConfig {
         if let Some(qt) = options.query_time {
             self.query_time = qt;
         }
+    }
+}
+
+pub(crate) struct AuthState {
+    pub(crate) cached_token: Option<TokenDetails>,
+    /// RSA10e/RSA10g: token params saved by authorize() for future renewals.
+    pub(crate) saved_token_params: Option<TokenParams>,
+    /// RSA10h: auth options saved by authorize(), replacing the client's
+    /// auth configuration for future renewals.
+    pub(crate) saved_auth_options: Option<AuthOptions>,
+    /// RSA10a: once authorize() has been called, token auth is used for all
+    /// subsequent requests even on a basic-auth (key) client.
+    pub(crate) forced_token_auth: bool,
+    /// RSA10k: cached offset between the server clock and the local clock,
+    /// captured when queryTime triggers a /time query.
+    pub(crate) time_offset_ms: Option<i64>,
+}
+
+/// Resolved Authorization header for a request. Basic vs Bearer matters
+/// beyond the header value: X-Ably-ClientId is only sent with basic auth
+/// (RSA7e2).
+#[derive(Clone, Debug)]
+pub(crate) enum AuthHeader {
+    Basic(String),
+    Bearer(String),
+}
+
+impl AuthHeader {
+    pub(crate) fn value(&self) -> String {
+        match self {
+            AuthHeader::Basic(v) => format!("Basic {}", v),
+            AuthHeader::Bearer(v) => format!("Bearer {}", v),
+        }
+    }
+}
+
+impl Rest {
+    /// The local clock adjusted by any cached server-time offset.
+    pub(crate) fn adjusted_now_ms(&self) -> i64 {
+        let offset = self
+            .inner
+            .auth_state
+            .lock()
+            .unwrap()
+            .time_offset_ms
+            .unwrap_or(0);
+        Utc::now().timestamp_millis() + offset
+    }
+
+    /// Invalidate the cached library token so the next acquisition renews it
+    /// (used by realtime token-error recovery, RTN14b/RTN15h2; called from
+    /// spawned connect tasks, never the connection loop).
+    pub(crate) fn invalidate_cached_token(&self) {
+        self.inner.auth_state.lock().unwrap().cached_token = None;
+    }
+
+    /// Resolve the auth configuration: the client's credential plus any
+    /// options stored by authorize() (RSA10h).
+    pub(crate) fn auth_config(&self) -> AuthConfig {
+        let opts = &self.inner.opts;
+        let mut cfg = AuthConfig {
+            key: None,
+            callback: None,
+            url: None,
+            token_request: None,
+            static_token: None,
+            method: opts
+                .auth_method
+                .clone()
+                .unwrap_or_else(|| "GET".to_string()),
+            headers: opts.auth_headers.clone(),
+            params: opts.auth_params.clone(),
+            query_time: opts.query_time,
+        };
+        match &opts.credential {
+            Credential::Key(k) => cfg.key = Some(k.clone()),
+            Credential::Callback(cb) => cfg.callback = Some(cb.clone()),
+            Credential::Url(u) => cfg.url = Some(u.clone()),
+            Credential::TokenRequest(tr) => cfg.token_request = Some(tr.clone()),
+            Credential::TokenDetails(_) => {}
+        }
+        let state = self.inner.auth_state.lock().unwrap();
+        if let Some(saved) = &state.saved_auth_options {
+            cfg.apply(saved);
+        }
+        cfg
+    }
+
+    /// Auth configuration with per-call AuthOptions applied on top.
+    pub(crate) fn auth_config_with(&self, options: Option<&AuthOptions>) -> AuthConfig {
+        let mut cfg = self.auth_config();
+        if let Some(o) = options {
+            cfg.apply(o);
+        }
+        cfg
+    }
+
+    /// Merge explicit token params with defaultTokenParams (RSA5c/RSA6c) and
+    /// the client's clientId (RSA7d).
+    pub(crate) fn effective_token_params(&self, params: Option<&TokenParams>) -> TokenParams {
+        let defaults = self.inner.opts.default_token_params.as_ref();
+        let p = params.cloned().unwrap_or_default();
+        TokenParams {
+            ttl: p.ttl.or_else(|| defaults.and_then(|d| d.ttl)),
+            capability: p
+                .capability
+                .or_else(|| defaults.and_then(|d| d.capability.clone())),
+            client_id: p
+                .client_id
+                .or_else(|| defaults.and_then(|d| d.client_id.clone()))
+                .or_else(|| self.inner.opts.client_id.clone()),
+            timestamp: p.timestamp,
+            nonce: p.nonce,
+        }
+    }
+
+    /// Obtain a token from the configured source. Precedence (RSA1): an
+    /// authCallback, then an authUrl, then a literal TokenRequest, then the
+    /// API key. Does not touch the cached library token.
+    pub(crate) async fn acquire_token(
+        &self,
+        params: &TokenParams,
+        cfg: &AuthConfig,
+    ) -> Result<TokenDetails> {
+        if let Some(cb) = &cfg.callback {
+            let token_result = cb.token(params).await.map_err(|e| {
+                let mut err = ErrorInfo::with_cause(
+                    ErrorCode::ErrorFromClientTokenCallback.code(),
+                    format!(
+                        "Auth callback error: {}",
+                        e.message.as_deref().unwrap_or("unknown")
+                    ),
+                    e,
+                );
+                err.status_code = Some(401);
+                err
+            })?;
+            // RSA4f: a token exceeding 128KiB is invalid output from the
+            // callback
+            const MAX_TOKEN_LENGTH: usize = 128 * 1024;
+            let oversized = |t: &str| t.len() > MAX_TOKEN_LENGTH;
+            return match token_result {
+                AuthToken::Details(td) if oversized(&td.token) => Err(ErrorInfo::with_status(
+                    ErrorCode::ClientConfiguredAuthenticationProviderRequestFailed.code(),
+                    401,
+                    "Token from authCallback exceeds the maximum token length",
+                )),
+                AuthToken::Token(s) if oversized(&s) => Err(ErrorInfo::with_status(
+                    ErrorCode::ClientConfiguredAuthenticationProviderRequestFailed.code(),
+                    401,
+                    "Token from authCallback exceeds the maximum token length",
+                )),
+                AuthToken::Details(td) => Ok(td),
+                AuthToken::Token(s) => Ok(TokenDetails::token(s)),
+                AuthToken::Request(tr) => self.exchange_token_request(&tr).await,
+            };
+        }
+
+        if let Some(url) = &cfg.url {
+            return self.fetch_token_from_url(url, params, cfg).await;
+        }
+
+        if let Some(tr) = &cfg.token_request {
+            return self.exchange_token_request(tr).await;
+        }
+
+        if let Some(key) = &cfg.key {
+            let timestamp = self.token_request_timestamp(params, cfg).await?;
+            let tr = key.sign_with_timestamp(params, timestamp)?;
+            return self.exchange_token_request(&tr).await;
+        }
+
+        if let Some(td) = &cfg.static_token {
+            return Ok(td.clone());
+        }
+
+        Err(ErrorInfo::with_status(
+            ErrorCode::NoWayToRenewAuthToken.code(),
+            401,
+            "No way to obtain or renew auth token",
+        ))
+    }
+
+    /// RSA8c: fetch a token from the authUrl. The TokenParams and authParams
+    /// are merged: appended as query params for GET (RSA8c1a), form-encoded
+    /// in the body for POST (RSA8c1b). The response is interpreted by
+    /// Content-Type: JSON is a TokenRequest (exchanged) or TokenDetails;
+    /// anything else is a literal token string.
+    async fn fetch_token_from_url(
+        &self,
+        auth_url: &str,
+        params: &TokenParams,
+        cfg: &AuthConfig,
+    ) -> Result<TokenDetails> {
+        let mut url = url::Url::parse(auth_url).map_err(|e| {
+            ErrorInfo::new(
+                ErrorCode::ErrorFromClientTokenCallback.code(),
+                format!("Invalid authUrl: {}", e),
+            )
+        })?;
+
+        // RSA8c1: merge TokenParams and authParams
+        let mut pairs: Vec<(String, String)> = Vec::new();
+        if let Some(ttl) = params.ttl {
+            pairs.push(("ttl".to_string(), ttl.to_string()));
+        }
+        if let Some(cap) = &params.capability {
+            pairs.push(("capability".to_string(), cap.clone()));
+        }
+        if let Some(cid) = &params.client_id {
+            pairs.push(("clientId".to_string(), cid.clone()));
+        }
+        if let Some(ts) = params.timestamp {
+            pairs.push(("timestamp".to_string(), ts.timestamp_millis().to_string()));
+        }
+        pairs.extend(cfg.params.iter().cloned());
+
+        let method = cfg.method.to_uppercase();
+        let mut headers = cfg.headers.clone();
+        let body = if method == "POST" {
+            headers.push((
+                "content-type".to_string(),
+                "application/x-www-form-urlencoded".to_string(),
+            ));
+            let encoded: String = pairs
+                .iter()
+                .map(|(k, v)| format!("{}={}", urlencoding::encode(k), urlencoding::encode(v)))
+                .collect::<Vec<_>>()
+                .join("&");
+            Some(encoded.into_bytes())
+        } else {
+            for (k, v) in &pairs {
+                url.query_pairs_mut().append_pair(k, v);
+            }
+            None
+        };
+
+        let req = HttpRequest {
+            method,
+            url: url.to_string(),
+            headers,
+            body,
+        };
+        let result = tokio::time::timeout(
+            self.inner.opts.http_request_timeout,
+            self.inner.http_client.execute(req),
+        )
+        .await;
+        let resp = match result {
+            Ok(Ok(resp)) => resp,
+            Ok(Err(e)) => {
+                return Err(ErrorInfo::with_status(
+                    ErrorCode::ErrorFromClientTokenCallback.code(),
+                    401,
+                    format!("authUrl request failed: {}", e),
+                ));
+            }
+            Err(_) => {
+                return Err(ErrorInfo::with_status(
+                    ErrorCode::ErrorFromClientTokenCallback.code(),
+                    401,
+                    "authUrl request timed out",
+                ));
+            }
+        };
+        if !(200..300).contains(&resp.status) {
+            return Err(ErrorInfo::with_status(
+                ErrorCode::ErrorFromClientTokenCallback.code(),
+                resp.status,
+                format!("authUrl returned status {}", resp.status),
+            ));
+        }
+
+        let ct = resp
+            .headers
+            .iter()
+            .find(|(k, _)| k.eq_ignore_ascii_case("content-type"))
+            .map(|(_, v)| v.as_str())
+            .unwrap_or("");
+        if ct.contains("application/json") {
+            let v: serde_json::Value = serde_json::from_slice(&resp.body)?;
+            if v.get("mac").is_some() && v.get("keyName").is_some() {
+                let tr: TokenRequest = serde_json::from_value(v)?;
+                self.exchange_token_request(&tr).await
+            } else {
+                Ok(serde_json::from_value(v)?)
+            }
+        } else {
+            // text/plain, application/jwt, or unspecified: a literal token
+            let token = String::from_utf8(resp.body).map_err(|e| {
+                ErrorInfo::new(
+                    ErrorCode::ErrorFromClientTokenCallback.code(),
+                    format!("Invalid token string from authUrl: {}", e),
+                )
+            })?;
+            Ok(TokenDetails::token(token))
+        }
+    }
+
+    /// RSA15: an obtained token's clientId must be compatible with the
+    /// client's configured clientId ("*" is compatible with anything).
+    pub(crate) fn check_client_id_compat(&self, td: &TokenDetails) -> Result<()> {
+        if let (Some(opt_cid), Some(tok_cid)) = (&self.inner.opts.client_id, &td.client_id) {
+            if tok_cid != "*" && tok_cid != opt_cid {
+                return Err(ErrorInfo::with_status(
+                    ErrorCode::IncompatibleCredentials.code(),
+                    401,
+                    format!(
+                        "Token clientId '{}' is incompatible with configured clientId '{}'",
+                        tok_cid, opt_cid
+                    ),
+                ));
+            }
+        }
+        Ok(())
+    }
+
+    /// Is token auth in effect (RSA4)? Token auth is triggered by
+    /// useTokenAuth, any token-bearing credential, or a previous authorize()
+    /// (RSA10a). Only a key-only client uses basic auth.
+    fn token_auth_in_effect(&self) -> bool {
+        if self.inner.opts.use_token_auth {
+            return true;
+        }
+        if self.inner.auth_state.lock().unwrap().forced_token_auth {
+            return true;
+        }
+        !matches!(self.inner.opts.credential, Credential::Key(_))
+    }
+
+    /// Resolve the Authorization header for a request: basic auth for a
+    /// key-only client (RSA2/RSA11), otherwise the cached token, with
+    /// pre-emptive renewal when it is known to be expired (RSA4b1) or a
+    /// client-side 40171 when it cannot be renewed (RSA4a2).
+    pub(crate) async fn get_auth_header(&self) -> Result<AuthHeader> {
+        if !self.token_auth_in_effect() {
+            if let Credential::Key(key) = &self.inner.opts.credential {
+                return Ok(AuthHeader::Basic(base64::encode(format!(
+                    "{}:{}",
+                    key.name, key.value
+                ))));
+            }
+        }
+
+        // Token auth
+        let cached = self.inner.auth_state.lock().unwrap().cached_token.clone();
+        let cfg = self.auth_config();
+        let can_renew = cfg.callback.is_some() || cfg.url.is_some() || cfg.key.is_some();
+        if let Some(td) = cached {
+            if !td.token.is_empty() {
+                if !td.is_expired(self.adjusted_now_ms()) {
+                    return Ok(AuthHeader::Bearer(td.token));
+                }
+                // RSA4a2: expired with no way to renew — fail client-side
+                if !can_renew {
+                    return Err(ErrorInfo::with_status(
+                        ErrorCode::NoWayToRenewAuthToken.code(),
+                        401,
+                        "Token expired and no way to renew it",
+                    ));
+                }
+            }
+        }
+
+        // Acquire a (new) library token using saved params (RSA10e) merged
+        // with defaults.
+        let saved = self
+            .inner
+            .auth_state
+            .lock()
+            .unwrap()
+            .saved_token_params
+            .clone();
+        let params = self.effective_token_params(saved.as_ref());
+        let td = self.acquire_token(&params, &cfg).await?;
+        self.check_client_id_compat(&td)?; // RSA15
+        self.inner.auth_state.lock().unwrap().cached_token = Some(td.clone());
+        Ok(AuthHeader::Bearer(td.token))
     }
 }
